Finalized Models

+++ migrations/20210615195114-create-item.js
@@ -15,7 +15,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       price: {
-        type: Sequelize.INTEGER
+        type: Sequelize.NUMERIC
       },
       status: {
         type: Sequelize.STRING
